Moved interfaces into own files
