+++ src/app/interfaces/vbs-task-interface.ts
@@ -1,0 +1,4 @@
+export interface VbsServiceCommunication {
+    statusTaskRemainingTime: string;
+    statusTaskInfoText: string;
+}
